Add "mixin class" to the "base-interface-final" proposal.
I went ahead and took the more aggressive step of also saying that you
can't mixin a class unless it has the "mixin" modifier. However, that
rule is gated by a language version so that it's not breaking.

Since "mixin class" allows you to preserve the previous behavior of
class+mixin, I think we can get away with doing this change in one step.
Basically, whenever you upgrade your library to the latest version, you
add "mixin" to whatever classes you want to support being used as mixins
and you're done.

If you think that's too aggressive, let me know and I can change this.
If we do that, though, it ends up meaning that the "mixin" modifier
doesn't actually *do* anything (except yell at you if you apply it to a
class with a generative constructor or non-Object superclass).