[patterns] Clarify which variables are valid in pattern assignments. (#2686)
This might be too specific. The real intent here is just that if it's
a local variable or parameter and you can assign to it outside of a
pattern, then you can assign to it in one too. If I correctly captured
those rules, great. If not, let me know and I'll tweak it.