Revise proposal.
I went through and fully incorporated the allowing mixin as a modifier
into the proposal. I specified the semantics for how modifiers on
mixin declarations come into effect.

And I went through *all* of the combinations, figured out which ones
made sense, and tried to come up with a rational set of allowed
combinations with well understood semantics for each.

I'm really sorry to do a giant rewrite right in the middle of a move
PR but since there were valid comments on the PR I just sort of went
for it.