Moved sealed-types and base-interface-final to accepted/ (#2722)
* Moved sealed-types and base-interface-final to accepted/

Also move competing packaged-libraries and type-modifiers proposals to
inactive/.

* Rename proposal to "type modifiers".

We decided to put this behind a separate experiment flag and we needed
a name for that flag. "base-interface-final-mixin" would be a mouthful,
so I resurrected the old "type modifiers" name (used for an earlier
superceded proposal) and renamed this proposal to match that.

* Revise proposal.

I went through and fully incorporated the allowing mixin as a modifier
into the proposal. I specified the semantics for how modifiers on
mixin declarations come into effect.

And I went through *all* of the combinations, figured out which ones
made sense, and tried to come up with a rational set of allowed
combinations with well understood semantics for each.

I'm really sorry to do a giant rewrite right in the middle of a move
PR but since there were valid comments on the PR I just sort of went
for it.

* Rename to "Class modifiers".

Also fix some links, add some context, and update the changelog.

* Remove unnecessary rule about mixing in final mixins.