Adjust `primitive operator ==` to `primitive equality`, and include `hashCode` (#2106)
This PR changes the notion of having a 'primitive operator `==`' to a similar notion of having 'primitive equality'. The latter specifies the same requirements for the operator `==`, but it adds the corresponding requirements for `hashCode`. As a result, we now know that an object with primitive equality allows for compile-time evaluation of its `hashCode` as well as its `==`.