Better exhaustiveness algorithm.
This one is based on Luc Maranget's "Warnings for pattern matching":

http://moscova.inria.fr/~maranget/papers/warn/warn.pdf

It passes all of the exhaustiveness tests in the prototype. It also
found a bug in the original prototype's algorithm.

Once you strip out all of the comments and stuff, it's only about 60
lines of code.

It's about 100x faster than the original prototype on the benchmark
(~10ms versus ~0.13ms on my machine).