[patterns] Try to clarify remark on non-"refutable" patterns that fail
I found I had to reread this paragraph a couple of times to understand
the point it was making.  (Possibly the logic error just fixed in it
can be taken as evidence that other readers found it a bit hard to
read, too.)  So here's an attempt at making it clearer.