[patterns] Slightly unpack logic for list length-checking
I found the preface "If `p` is empty or has any non-rest elements",
and its relationship with the sub-cases it encloses, required some
re-reading to put together.  It's meant as a way of saying "If `p` is
not `[...]`", as referred to in the rationale line that follows it;
but to see that, one has to work through a bit of Boolean logic to
invert the condition.

So here's a rewording which I hope is clearer, by un-inverting the
condition.

The usage "then do nothing for checking the length" is perhaps a bit
unconventional in spec-ese.  But it's parallel to the "then do nothing
for checking this element" found in the head and tail cases below.