[patterns] Clarify when a set of map patterns is exhaustive
For lists, one can use the empty list pattern `[]` to make an
exhaustive set of list patterns.

But for maps, the only way to make an exhaustive set of map patterns
is the trivial way, using the pattern `{...}` that's exhaustive (on
maps) all by itself.

So move the remark about exhaustive sets of map patterns out of the
context of the empty map pattern `{}`, and up to the discussion of
`{...}`.  Also say that these are the only exhaustive sets of map
patterns, and why.

Fix #2818.