Change `==` in constant expressions to use primitive equality (#2874)
Change `==` in constant expressions to use primitive equality.

Note that the left operand of `==` can also be an instance of `double`. They don't have primitive equality, but we did not want to introduce a breaking change, and `1.5 == 2.5` was a correct constant expression until now, so we keep it that way as a special exception.