Specify not unsound (#2883)
Reword the rationale paragraph about Dart typing and soundness. The new text indicates that Dart has soundness properties (the old text just said that it was 'unsound', which is unfair). It also indicates that the evolution of Dart has been, and likely will be, in the direction of more static type safety.