Update augmentation library spec for class modifiers (#2917)
I removed the line about `abstract` and instead chose to say that all class modifiers must be present on both the augmentation type and the original type. This is I think consistent with the requirement to duplicate the type parameters, and it ensures people won't be confused about the modifiers on a class when looking at only the augmentation.

Open for debate certainly though! We could go the opposite direction and drop the `class/mixin/enum` keywords from augmentation libraries, as well as duplicate type parameters, thus eliminating all redundant information.