Update arguments spec to allow TypeAnnotations instead of Identifiers (#3146)
The real intention of allowing Identifiers was to allow types, but we also want to allow type arguments. This removes the ability to have arbitrary Identifier arguments and replaces that with TypeAnnotation arguments.

We might want to add Identifier arguments also, to handle non-type identifiers, but I don't really have a use case.