Extension types feature specification (#3182)
Modify the inline class feature specification to specify extension types. This is a feature renaming, but also a reintroduction of a special case of the primary constructor syntax, and an enhancement with support for non-extension type superinterfaces (cf. #3090). Note that many types cannot be used as superinterfaces (including `Function` and function types, records, `T?` for any `T`); we may choose to enable a broader set of types in the future, but at this time we prefer to keep it simple. For instance, `UP` may need to be revised if we accept a broader set of types.