add basic checks extension macro (#3315)
This macro generates extensions for a list of types, for use with package:checks. It isn't yet runnable as I need to do a bit of upstream work, but should be a good start.

I also updated the spec for extensions, allowing introspection on the on type as well as all generic type parameters.