Add error for non-covariant type variable in superinterface (#3412)
Add compile-time error for non-covariant type variable in superinterface, just like the rule we have for classes, mixins, etc. It may be sound to allow this, but for now we prefer the simplicity and consistency of having the same rule everywhere. Also, the ability to violate this constraint does not seem to allow for anything useful.