Adjust the rule about await e where the type of e is an extension type (#3560)
Adjust the rule about a compile-time error that may occur at `await e` when the type of e is an extension type. Previously the error only occurred when the static type of `e` was an extension type that does not implement `Future`. 

This PR introduces the concept of a type being _incompatible with await_ and specifies that this error occurs whenever the static type of `e` is incompatible with await. With this, the error occurs in several additional cases: When the static type of `e` is a type variable whose bound is such an extension type, or a nullable type whose operand is such an extension type, etc., recursively.