Augmentation grammar: Fix libraryName, reorder two top level alternatives, remove muthedSignature rule (#3667)
Adjust the augmentation grammar slightly. We used to have this:

```ebnf
libraryName ::= metadata 'augment'? 'library' ( dottedIdentifierList? |  uri ) ';'
```

but that allows for things like `library 'package:whatever/whatever.dart';` which is not useful. This PR adjusts the rule to allow the `uri` with an `augment library`, and not with a plain `library`.

This PR also removes the rule for `methodSignature` because it only changed by having `'augment'?` in some alternatives: `classMemberDeclaration` already has `'augment'?` in front of `methodSignature`, and we don't need to allow two occurrences of `augment` on the same declaration.

Finally, this PR swaps the alternatives for `extensionDeclaration` and `extensionTypeDeclaration` in the `topLevelDeclaration` rule, to match the approach taken in `Dart.g`.