Adjust the extensionDeclaration syntax (#3701)
In response to https://github.com/dart-lang/language/issues/3687, this PR changes the grammar rule `<extensionDeclaration>` such that an augmenting extension declaration does not have an `on` clause syntactically. The compile-time error caused by having such an `on` clause is removed (it is now a syntax error). The text about extensions is adjusted slightly elsewhere to fit these changes.

This resolves a contradiction: An `augment extension` declaration used to be a syntax error if it did not have an `on` clause, but an `augment extension` declaration _with_ an `on` clause was a compile-time error. The new syntax also enforces that an `augment extension` declaration cannot be anonymous, which is desirable because augmentation of an anonymous extension is rather ill defined (for example, if we insist that each anonymous extension has a fresh private name then there will never be a match on the name and hence the `augment extension` declaration will always be an error because it doesn't match any original declaration).