Add a proposal about inference of the modifier `required` (#3779)
Add a feature specification about a piece of syntactic sugar that implicitly induces the modifier `required` in the cases where the corresponding named formal parameter has a potentially non-nullable type (which means that it is an error to omit `required`). Note that the mechanism is only applicable to the formal parameter lists of concrete functions, it can not be used with abstract instance method declarations, and it can not be used with function types.