Update the primary constructors proposal (#3836)
Make adjustments to the proposal about primary constructors such that it is in line with the results from discussions at the recent physical language team meeting.

In particular, remove support for a `primary` constructor which is declared in the body of the class/enum/etc, and only allow for the form which is part of the header of the class declaration. Also remove the ability to infer the modifier `required` in a formal parameter declaration.

(Inferred `required` is now a separate proposal, see https://github.com/dart-lang/language/blob/main/working/0015-infer-required/feature-specification.md.)

The ability to omit `final` from every parameter when the primary constructor is constant is preserved because we already have the ability to omit `final` in the primary constructor of an extension type declaration.