Consistently use terminology "instance implementing" and "dynamic type".
In the code review discussion
https://github.com/dart-lang/language/pull/3850#discussion_r1635527576,
we agreed to use the terminology "instance satisfying type T" to mean
"instance whose runtime type is T or a subtype thereof", and I made
the necessary changes to
`resources/type-system/inference.md`. However, after digging through
`dartLangSpec.tex`, it appears that the terminology we were already
using in the spec for this concept is "instance implementing type
T". This commit modifies `resources/type-system/inference.md` to
match the spec terminology.

It also fixes two places in the spec where we were incorrectly saying
"instance of type T", but we meant "instance implementing type T".

Furthermore, the spec consistently uses the term "dynamic type of `v`"
for the concept that we informally refer to as the "runtime type of
`v`", so this commit updates `resources/type-system/inference.md` to
follow the spec convention.