Simplify reasoning about await soundness using a lemma about flatten. (#3920)
The lemma is that for all types `T`, `T <: FutureOr<flatten(T)>`. The
lemma was claimed (without proof) in `dartLangSpec.tex`, so I've added
a (non-normative) proof of it to the spec, right after the claim.

This made it possible to significantly simplify the reasoning around
the soundness of `await` in `inference.md`.

Thanks to @lrhn for the core idea behind this.