Fix typo in comment about `augmented` (#3950)
Fix a typo in a comment about `augmented`, that it is a normal identifier outside an augmenting declaration, not an augmented one. Also fix another typo: we cannot have a setter whose name is `augmented=` in an augmenting declaration (the old text forgot about `=` and only restricted declarations named `augmented`).