Slightly clarify and copy-edit the augmentation ordering. (#3981)
Slightly clarify and copy-edit the augmentation ordering.

There are no semantic changes here, I just found the prose a little
unclear on first read. In particular:

```
*   If the `part` directive including the file that contains *A*
    is syntactically before the `part` directive including the file that
    contains *B* in source order, then *A* is before *B*.
```

Is ambiguous because "includes" is transitive, so there may be multiple
files whose `part` directives include A or B and it's not clear what
that means when those `part` directives aren't in the same file.

Also made a couple of style and copy tweaks while I was at it.

I like the semantics!