Eliminate the notion of "bound resolution" from inference.md.
In its place, the spec's definition of "T_0 bounded" is tightened up
so that for any type T, there is exactly one type T_0 such that T is
T_0 bounded. Then, instead of saying "let U_0 be the bound resolution
of T_0", we can say "let U_0 be the unique type such that T_0 is U_0
bounded."

This avoids some redundancy, since we no longer need to define
separate but related notions of "T_0 bounded" and "bound resolution".