Clarify extension member resolution (#4116)
Clarify extension member resolution: The applicability of a given extension for an invocation of an extension instance member `m` would require that the static type of the syntactic receiver does not have 'a member whose basename is the basename of `m`'. This PR makes it explicit that this is 'an instance member whose basename is the basename of `m`', and it isn't relevant if the declaration that induced said type has a static member with such a name.