Adjust the constant expression rules for lists, maps, and conditional expressions (#4313)
Adjust the constant expression rules for lists, sets, maps, constant object expressions, and conditional expressions. For the latter, it is specified that a conditional expression which is not potentially constant is also not constant. For the others, the rules in section 'Constants' are simplified (it's redundant with the same rule as stated in other sections).

There is no breaking change and no implementation effort, because the newly specified behavior is the already implemented behavior (for the analyzer as well as the CFE).