Create a script to simplify the language specification (#4340)
This PR adds the script 'specification/scripts/simplify_specification.dart' which will read 'dartLangSpec.tex', remove LaTeX comments, remove commentary and rationale, and join the lines of paragraphs to a single line. The result is saved as 'dartLangSpec-terse.tex' which is more convenient for searching (grepping) and for reading normative text.

Moreover, whitespace in 'dartLangSpec.tex' is modified for improved consistency. In particular, commentary and rationale is indented consistently and a number of line breaks have been moved a bit in order to avoid lines longer than 80 characters.