[flow analysis] Clean up promoted types ordering and use of `::` (#4369)
The `promotedTypes` list in a `VariableModel` is now consistently
ordered from least promoted type to most promoted type, matching the
list used in the implementation.

Also, the use of `top::rest` to refer to a stack is changed to
`[...rest, top]`. This is consistent with the way some stacks are
represented in the flow analysis implementation (as a list where
pushing and popping happens at the end of the list).

Fixes https://github.com/dart-lang/language/issues/4367.