Add a feature specification for the metaobject proposal (#4301)
This PR adds a feature specification proposal for the 'metaobjects' feature which is also the core contents of issue #4200. The basic idea is that the step that yields a reified type (an instance of `Type`) from a given type can be customized to yield an instance with additional superinterfaces, and this will allow us to perform actions based on a type, including invocation of constructors and static members. For instance, `T(42)` could be used to invoke a constructor of type `C` if `T` is a type variable whose value is a class `C` and the class `C` has a constructor that accepts this actual argument list.