Add exception about extension members in error about nullable receivers (#4318)
Add exception about extension members in the error about nullable receivers (`e.m()` can be OK even when `e` has type `int?` if `m` is an extension method with an on-type that `int?` can match). Also mention that extension type members need to be an exception (an extension type that doesn't implement `Object`, directly or indirectly, is a potentially nullable type, and we shouldn't report all invocations of its members as errors).