Clarify the `void` allowlist wrt `[]` and `[]=` (#4454)
This PR adds a clarification to the `void` allowlist, specifying explicitly that `e1` can have type `void` in the case where `e0[e1]` is encountered and the statically known operator `[]` of `e1` has a parameter whose type is `void`, and similarly for `e0[e1] = e2` where both `e1` and `e2` can have type `void` when the corresponding formal parameters have type `void`. It also clarifies that the same rule applies to setters. Finally, it clarifies the item about local variable / parameter assignments.

This PR also changes the word 'parameter' to 'argument' in the same item, in a location where it is clearly intended to refer to an actual argument, not a formal parameter.

See also https://github.com/dart-lang/sdk/issues/57071.