[flow analysis] Rework control flow statements and expressions. (#4427)
This change adjusts the specification of flow analysis for the
following kinds of statements and expressions, to match what was
actually implemented:

- conditional expressions
- if-null expressions
- logical and/or expressions
- null check operators
- break statements
- continue statements
- return statements
- if statements
- while statements
- for statements
- do while statements
- for each statements
- switch statements
- try statements

Also, the paragraphs describing "Local variable conditional
assignment" and "Conditional assignment to a non local-variable" are
dropped. These paragraphs were quite inaccurate, and the necessary
infrastructure isn't in place yet to replace them with more accurate
descriptions.

Finally, there's a minor fix: the definition of a promotion chain was
backwards (types are listed in least-promoted to most-promoted order).