Move "Parts with Imports" to a separate proposal. (#4469)
In #3939, there seemed to be consensus that "parts with imports" should
be its own proposal with its own experiment flag. This PR does that.
While I was at it, I renamed "Augmentation libraries" to just
"Augmentations" (since it's not a new kind of library anymore) and gave
that an experiment flag too.

There are no changes to the text of either proposal. (Well, I did get
rid of smart quotes in the parts with imports one...)

Fix #3939.